Handle case of too many games
Att m2x plans are done by device, so a 250 device plan will eventually
not be enough as games rise an fall in popularity (really this is the
long tail, the top 25 games don't really change)

Also refactored some of the create/get device stuff. Added new config
settings and documented in the readme.